save codes

+++ src/main/java/br/com/centroinfo/api/dtos/saleDTO/NotaDTO.java
@@ -1,5 +1,6 @@
 package br.com.centroinfo.api.dtos.saleDTO;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -14,14 +15,14 @@ public class NotaDTO {
     private String personName;
     private String street;
     private List<ItemNotaDTO> items;
-    private Double totalSale;
-    private Double discount;
-    private Double totalNote;
+    private BigDecimal totalSale;
+    private BigDecimal discount;
+    private BigDecimal totalNote;
 
     public NotaDTO(Long saleId, LocalDateTime issueDate,
     String branchName, String userName, String street,
-    String personName, List<ItemNotaDTO> items, Double totalSale,
-    Double discount, Double totalNote) {
+    String personName, List<ItemNotaDTO> items, BigDecimal totalSale,
+    BigDecimal discount, BigDecimal totalNote) {
         this.saleId = saleId;
         this.issueDate = issueDate;
         this.branchName = branchName;
