+++ src/main/java/br/com/centroinfo/api/entities/sales/Sale.java
@@ -10,6 +10,8 @@ import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
+
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -43,11 +45,11 @@ public class Sale {
     @JoinColumn(name = "operation_sale_id")
     private OperationSale operationSale;
     @Column(name = "total_sale")
-    private Double totalSale;
+    private BigDecimal totalSale;
     @Column(name = "discount")
-    private Double discount;
+    private BigDecimal discount;
     @Column(name = "total_note")
-    private Double totalNote;
+    private BigDecimal totalNote;
     @OneToMany(mappedBy = "sale", cascade = CascadeType.ALL)
     private List<ItemSale> itemsSale;
     @OneToMany(mappedBy = "sale" , cascade = CascadeType.ALL)
