+++ src/main/java/br/com/centroinfo/api/services/sale/NotaService.java
@@ -4,6 +4,7 @@ import br.com.centroinfo.api.dtos.saleDTO.ItemNotaDTO;
 import br.com.centroinfo.api.dtos.saleDTO.NotaDTO;
 import br.com.centroinfo.api.entities.sales.Sale;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
@@ -15,8 +16,8 @@ public class NotaService {
         List<ItemNotaDTO> itemDTOs = sale.getItemsSale().stream().map(item -> {
             Long id = item.getItem().getId();
             String item_ = item.getItem().getName();
-            Integer amount = item.getAmount();
-            Double price = item.getPrice();
+            BigDecimal amount = item.getAmount();
+            BigDecimal price = item.getPrice();
             String barCode = item.getItem().getBarCode();
             String sector = item.getItem().getSubGroup().getName();
             String brand = item.getItem().getBrand().getName();
